requirements-dev.txt
black==24.10.0
isort==5.13.2
flake8==7.1.1
pytest==8.3.3
pytest-django==4.9.0
pre-commit==3.8.0
